detail chages: modularization, error handling etc
