+++ src/components/Directions/Directions.css
@@ -1,16 +1,15 @@
 .directions-container {
-    height: 100vh; /* Full viewport height */
-    width: 100%;   /* Full width */
+    height: 100vh;
+    width: 100%;
     position: fixed;
 }
 
 .map {
-    height: 100%; /* Map should fill the container */
+    height: 100%;
     width: 100%;
-
 }
 
-.directions{
+.directions {
     position: relative;
     z-index: 1;
     top: -95vh;
@@ -22,7 +21,7 @@
     flex-direction: column;
     padding: 1.5rem;
 
-    button{
+    button {
         background-color: transparent;
         color: white;
         text-decoration: underline;
