+++ src/components/LoadingPage/LoadingPage.css
@@ -1,4 +1,4 @@
-#loading{
+#loading {
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -7,7 +7,7 @@
     background-color: var(--main-blue-ligth);
     height: 100vh;
 
-    img{
+    img {
         width: 30vw;
     }
 }
