+++ src/components/RegisterPage/RegisterPage.css
@@ -1,20 +1,19 @@
-#registerPage{
+#registerPage {
     display: flex;
     background-color: var(--main-blue-ligth);
     height: 90vh;
-    display: flex;
     justify-content: center;
     align-items: center;
     color: var(--blue-dark);
 }
 
-.promptContainer{
+.promptContainer {
     width: 80vw;
     height: 80vh;
     border: none;
 }
 
-.promptContainerImg{
+.promptContainerImg {
     color: white;
     display: flex;
     flex-direction: column;
@@ -22,39 +21,27 @@
     justify-content: center;
     border-radius: 32px 0px 0px 32px;
     gap: 0.5rem;
-
-    img{
-        width: fit-content;
-    }
-
-    h2{
-        font-weight: 400;
-    }
-
-    p{
-        font-weight: 300;
-    }
 }
 
-#promptContainerImgRegister{
+#promptContainerImgRegister {
     background-color: var(--blue-detail);
 }
 
-.prompt{
+.prompt {
     align-items: center;
 
-    form{
+    form {
         display: flex;
         align-items: center;
         justify-content: center;
 
-        input{
+        input {
             width: 80%;
         }
     }
 }
 
-#registerButton{
+#registerButton {
     background-color: var(--blue-dark);
     color: white;
 }
