+++ src/components/ResultPage/ResultPage.css
@@ -1,4 +1,4 @@
-.resultPage{
+.resultPage {
     display: flex;
     flex-direction: column;
     text-align: center;
@@ -7,7 +7,7 @@
     flex-grow: 4;
 }
 
-.resultBlock{
+.resultBlock {
     display: flex;
     flex-wrap: wrap;
     overflow: hidden;
@@ -31,7 +31,7 @@
     flex-direction: column;
     align-items: center;
 
-    .columnHeader{
+    .columnHeader {
         height: 10vh;
         align-content: center;
         justify-content: center;
@@ -42,13 +42,13 @@
         background-color: black;
         border-radius: 29px 29px 0px 0px;
 
-        h3{
+        h3 {
             width: fit-content;
             height: fit-content;
         }
     }
 
-    button{
+    button {
         background-color: var(--main-blue-ligth);
         padding: 10px;
         border: solid black;
@@ -85,7 +85,8 @@
     justify-content: center;
     align-items: center;
     display: flex;
-    h1{
+
+    h1 {
         background-color: rgba(255, 255, 255, 0.83);
         border-radius: 32px;
         padding: 1%;
@@ -93,5 +94,4 @@
         width: fit-content;
         height: fit-content;
     }
-  }
-
+}
