+++ src/components/ResultPage/ResultPage.tsx
@@ -1,29 +1,22 @@
 import { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
+import { State } from '../../types/types';
 import { DragDropContext, Droppable, Draggable, DraggableLocation } from "react-beautiful-dnd";
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import _ from "lodash";
 import './ResultPage.css'
 
 function Result() {
-    interface State {
-        [key: string]: {
-            title: string;
-            places: { id: number; name: string; description: string }[];
-        };
-    }
-
-    interface Errors {
-        prompt?: string;
-        // Add other error properties if needed
-    }
-
     const location = useLocation();
     const navigate = useNavigate();
+
     const itinerary = typeof location.state.response === 'string'
         ? JSON.parse(location.state.response)
         : location.state.response
     const destinationCity = location.state && location.state.city;
     const hasResponseData = itinerary && Object.keys(itinerary).length > 0;
+
     const initialState: State = hasResponseData
         ? Object.entries(itinerary).reduce((acc, [key, value]) => {
             if (key.startsWith('day')) {
@@ -37,7 +30,6 @@ function Result() {
         : {};
 
     const [state, setState] = useState<State>(initialState);
-    const [errors, setErrors] = useState<Errors>({});
 
     const handleDragEnd = ({ destination, source }: { destination: any, source: DraggableLocation }) => {
         if (!destination) {
@@ -65,12 +57,11 @@ function Result() {
         const places = state[dayKey].places.map((place) => place.name);
 
         if (places.length < 2) {
-            setErrors((e) => ({ ...e, prompt: "At least two places are required to generate directions." }));
+            toast.error("At least two places are required to generate directions.");
             return;
         }
 
-        navigate('/routeDirections', { state: { places: places, destinationCity: destinationCity  } })
-
+        navigate('/routeDirections', { state: { places: places, destinationCity: destinationCity } })
     };
 
     return (
@@ -78,6 +69,8 @@ function Result() {
             <header className='background'>
                 <h1>Here is your itinerary, to {destinationCity}</h1>
             </header>
+
+            <ToastContainer position="top-right" autoClose={6000} hideProgressBar={false} />
 
             {!hasResponseData ? (
                 <div>No itinerary available. Please check back later or try a different destination.</div>
