+++ src/components/UserPage/UserPage.tsx
@@ -2,7 +2,7 @@ import './UserPage.css'
 import cardImage from '../../assets/cardImage.svg'
 
 export default function UserPage() {
-    return(
+    return (
         <div id='userpage'>
             <header>
                 <h1>Welcome Back!</h1>
@@ -13,12 +13,12 @@ export default function UserPage() {
                     <div className='card'>
                         <img src={cardImage} alt="" />
                         <h4>Trip to London</h4>
-                        <h5>14 - 29 June | Solo Trip <br/> 15 day trip</h5>
+                        <h5>14 - 29 June | Solo Trip <br /> 15 day trip</h5>
                     </div>
                     <div className='card'>
                         <img src={cardImage} alt="" />
                         <h4>Trip to London</h4>
-                        <h5>14 - 29 June | Solo Trip <br/> 15 day trip</h5>
+                        <h5>14 - 29 June | Solo Trip <br /> 15 day trip</h5>
                     </div>
                 </div>
             </div>
