+++ src/index.css
@@ -3,7 +3,7 @@
   padding: 0;
   box-sizing: border-box;
   font-family: 'Rubik';
-  /* height: 100%; */
+  font-weight: 500;
   width: 100%;
 
   button{
@@ -28,8 +28,15 @@
   h2{
     font-size: 2.5rem;
   }
-}
 
+  p{
+    font-weight: 300;
+  }
+
+  img{
+    width: fit-content;
+  }
+}
 
 :root{
   --main-blue-ligth: rgba(167, 195, 221, 1);
@@ -37,3 +44,10 @@
   --blue-detail: #8CC1FF;
   --text-gray: #5E6282;
 }
+
+.bordered-element {
+  border-radius: 10px;
+  border: solid black;
+  border-width: 1px 4px 4px 1px;
+}
+
